requirements.txt
# Hospital Management System - Dependencies

# Web Framework
Flask==3.0.0
flask-cors==4.0.0

# Authentication
PyJWT==2.8.0

# Notifications
twilio==9.8.6

# HTTP Requests
requests==2.32.3
